Recompile project
